+++ Project/ServerApp/client
@@ -1,1 +1,1 @@
-/Users/jlu/temp/fullstack hackathon/Project/ClientApp/build
+/Users/andrew/fullstack-hackathon/Project/ClientApp/build
